+++ packages/core/src/view/undoable_changes/CellAttributeChange.ts
@@ -14,9 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { isNullish } from '../../util/Utils';
+import { isNullish } from '../../internal/utils';
 import Cell from '../cell/Cell';
-
 import type { UndoableChange } from '../../types';
 
 /**